chore: rename functions to utils
