+++ src/common/utils/replaceSpace.ts
